requirements.txt
Mastodon.py==1.8.1
beautifulsoup4==4.12.2
